+++ linter/linter.go
@@ -48,6 +48,8 @@ func (l *Linter) Lint(node ast.Node, ctx *context.Context, isMain bool) types.Ty
 		l.lintUnusedBackends(ctx)
 		l.lintUnusedSubroutines(ctx)
 		l.lintUnusedGotos(ctx)
+		l.lintUnusedPenaltyboxes(ctx)
+		l.lintUnusedRatecounters(ctx)
 	}
 
 	return types.NeverType
@@ -94,6 +96,24 @@ func (l *Linter) lintUnusedSubroutines(ctx *context.Context) {
 			continue
 		}
 		l.Error(UnusedDeclaration(s.Decl.GetMeta(), s.Decl.Name.Value, "subroutine").Match(UNUSED_DECLARATION))
+	}
+}
+
+func (l *Linter) lintUnusedPenaltyboxes(ctx *context.Context) {
+	for _, p := range ctx.Penaltyboxes {
+		if p.IsUsed {
+			continue
+		}
+		l.Error(UnusedDeclaration(p.Decl.GetMeta(), p.Decl.Name.Value, "penaltybox").Match(UNUSED_DECLARATION))
+	}
+}
+
+func (l *Linter) lintUnusedRatecounters(ctx *context.Context) {
+	for _, rc := range ctx.Ratecounters {
+		if rc.IsUsed {
+			continue
+		}
+		l.Error(UnusedDeclaration(rc.Decl.GetMeta(), rc.Decl.Name.Value, "ratecounter").Match(UNUSED_DECLARATION))
 	}
 }
 
@@ -228,7 +248,7 @@ func (l *Linter) lintVCL(vcl *ast.VCL, ctx *context.Context) types.Type {
 	return types.NeverType
 }
 
-//nolint:gocognit
+//nolint:gocognit,funlen
 func (l *Linter) factoryRootStatements(vcl *ast.VCL, ctx *context.Context) []ast.Statement {
 	var statements []ast.Statement
 	for _, stmt := range vcl.Statements {
@@ -298,17 +318,48 @@ func (l *Linter) factoryRootStatements(vcl *ast.VCL, ctx *context.Context) []ast
 			}
 			statements = append(statements, stmt)
 		case *ast.SubroutineDeclaration:
-			if err := ctx.AddSubroutine(t.Name.Value, &types.Subroutine{Decl: t, Body: t.Block}); err != nil {
-				e := &LintError{
-					Severity: ERROR,
-					Token:    t.Name.GetMeta().Token,
-					Message:  err.Error(),
+			if t.ReturnType != nil {
+				if context.IsFastlySubroutine(t.Name.Value) {
+					err := &LintError{
+						Severity: ERROR,
+						Token:    t.ReturnType.GetMeta().Token,
+						Message:  fmt.Sprintf("State-machine method %s may not have a return type", t.Name.Value),
+					}
+					l.Error(err.Match(SUBROUTINE_INVALID_RETURN_TYPE))
 				}
-				l.Error(e.Match(SUBROUTINE_DUPLICATED))
+
+				returnType, ok := ValueTypeMap[t.ReturnType.Value]
+				if !ok {
+					err := &LintError{
+						Severity: ERROR,
+						Token:    t.ReturnType.GetMeta().Token,
+						Message:  fmt.Sprintf("Unexpected variable type found: %s", t.ReturnType.Value),
+					}
+					l.Error(err.Match(SUBROUTINE_INVALID_RETURN_TYPE))
+				}
+
+				if err := ctx.AddUserDefinedFunction(t.Name.Value, getSubroutineCallScope(t), returnType); err != nil {
+					err := &LintError{
+						Severity: ERROR,
+						Token:    t.Name.GetMeta().Token,
+						Message:  err.Error(),
+					}
+					l.Error(err.Match(SUBROUTINE_DUPLICATED))
+				}
+			} else {
+				err := ctx.AddSubroutine(t.Name.Value, &types.Subroutine{Decl: t, Body: t.Block})
+				if err != nil {
+					e := &LintError{
+						Severity: ERROR,
+						Token:    t.Name.GetMeta().Token,
+						Message:  err.Error(),
+					}
+					l.Error(e.Match(SUBROUTINE_DUPLICATED))
+				}
 			}
 			statements = append(statements, stmt)
 		case *ast.PenaltyboxDeclaration:
-			if err := ctx.AddPenaltybox(t.Name.Value, &types.Penaltybox{Penaltybox: t}); err != nil {
+			if err := ctx.AddPenaltybox(t.Name.Value, &types.Penaltybox{Decl: t}); err != nil {
 				e := &LintError{
 					Severity: ERROR,
 					Token:    t.Name.GetMeta().Token,
@@ -318,7 +369,7 @@ func (l *Linter) factoryRootStatements(vcl *ast.VCL, ctx *context.Context) []ast
 			}
 			statements = append(statements, stmt)
 		case *ast.RatecounterDeclaration:
-			if err := ctx.AddRatecounter(t.Name.Value, &types.Ratecounter{Ratecounter: t}); err != nil {
+			if err := ctx.AddRatecounter(t.Name.Value, &types.Ratecounter{Decl: t}); err != nil {
 				e := &LintError{
 					Severity: ERROR,
 					Token:    t.Name.GetMeta().Token,
@@ -400,6 +451,17 @@ func (l *Linter) lintBackendProperty(prop *ast.BackendProperty, ctx *context.Con
 				l.Error(InvalidType(v.Value.GetMeta(), v.Key.Value, kt, vt).Match(BACKEND_SYNTAX))
 			}
 		}
+
+		err := isProbeMakingTheBackendStartAsUnhealthy(*t)
+		if err != nil {
+			err := &LintError{
+				Severity: WARNING,
+				Token:    prop.Key.GetMeta().Token,
+				Message:  err.Error(),
+			}
+			l.Error(err.Match(BACKEND_PROBER_CONFIGURATION))
+		}
+
 	default:
 		// Otherwise, simply compare key type
 		kt, ok := BackendPropertyTypes[prop.Key.Value]
@@ -603,7 +665,15 @@ func (l *Linter) lintSubRoutineDeclaration(decl *ast.SubroutineDeclaration, ctx 
 		l.Error(InvalidName(decl.Name.GetMeta(), decl.Name.Value, "sub").Match(SUBROUTINE_SYNTAX))
 	}
 
-	cc := ctx.Scope(getSubroutineCallScope(decl))
+	scope := getSubroutineCallScope(decl)
+	var cc *context.Context
+	if decl.ReturnType != nil {
+		returnType := ValueTypeMap[decl.ReturnType.Value]
+		cc = ctx.UserDefinedFunctionScope(decl.Name.Value, scope, returnType)
+	} else {
+		cc = ctx.Scope(scope)
+	}
+
 	// Switch context mode which corredponds to call scope and restore after linting block statements
 	defer func() {
 		// Lint declared variables are used
@@ -611,6 +681,9 @@ func (l *Linter) lintSubRoutineDeclaration(decl *ast.SubroutineDeclaration, ctx 
 		cc.Restore()
 	}()
 	l.lint(decl.Block, cc)
+	// We are done linting inside the previous scope so
+	// we dont need the return type anymore
+	cc.ReturnType = nil
 
 	// If fastly reserved subroutine name (e.g vcl_recv, vcl_fetch, etc),
 	// validate fastly specific boilerplate macro is embedded like "FASTLY recv"
@@ -1013,6 +1086,49 @@ func (l *Linter) lintLogStatement(stmt *ast.LogStatement, ctx *context.Context) 
 }
 
 func (l *Linter) lintReturnStatement(stmt *ast.ReturnStatement, ctx *context.Context) types.Type {
+	if ctx.ReturnType != nil {
+		if stmt.HasParenthesis {
+			l.Error(&LintError{
+				Severity: ERROR,
+				Token:    stmt.Token,
+				Message:  fmt.Sprintf("function %s: only actions should be enclosed in ()", ctx.CurrentFunction()),
+			})
+		}
+
+		if stmt.ReturnExpression == nil {
+			lintErr := &LintError{
+				Severity: ERROR,
+				Token:    stmt.Token,
+				Message:  fmt.Sprintf("function %s must have a return value", ctx.CurrentFunction()),
+			}
+			l.Error(lintErr.Match(SUBROUTINE_INVALID_RETURN_TYPE))
+			return types.NeverType
+		}
+
+		err := isValidReturnExpression(*stmt.ReturnExpression)
+		if err != nil {
+			lintErr := &LintError{
+				Severity: ERROR,
+				Token:    (*stmt.ReturnExpression).GetMeta().Token,
+				Message:  err.Error(),
+			}
+			l.Error(lintErr.Match(SUBROUTINE_INVALID_RETURN_TYPE))
+			return types.NeverType
+		}
+
+		cc := l.lint(*stmt.ReturnExpression, ctx)
+		// Condition expression return type must be BOOL or STRING
+		if !expectType(cc, *ctx.ReturnType) {
+			lintErr := &LintError{
+				Severity: ERROR,
+				Token:    (*stmt.ReturnExpression).GetMeta().Token,
+				Message:  fmt.Sprintf("function %s return type is incompatible with type %s", ctx.CurrentFunction(), cc.String()),
+			}
+			l.Error(lintErr.Match(SUBROUTINE_INVALID_RETURN_TYPE))
+		}
+		return types.NeverType
+	}
+
 	// legal return actions are different in subroutine.
 	// https://developer.fastly.com/learning/vcl/using/#the-vcl-request-lifecycle
 	expects := make([]string, 0, 3)
@@ -1048,13 +1164,13 @@ func (l *Linter) lintReturnStatement(stmt *ast.ReturnStatement, ctx *context.Con
 	}
 
 	// return statement may not have arguments, then stop linting
-	if stmt.Ident == nil {
+	if stmt.ReturnExpression == nil {
 		return types.NeverType
 	}
 
-	if !expectState(stmt.Ident.Value, expects...) {
+	if !expectState((*stmt.ReturnExpression).String(), expects...) {
 		l.Error(InvalidReturnState(
-			stmt.Ident.GetMeta(), context.ScopeString(ctx.Mode()), stmt.Ident.Value, expects...,
+			(*stmt.ReturnExpression).GetMeta(), context.ScopeString(ctx.Mode()), (*stmt.ReturnExpression).String(), expects...,
 		).Match(RESTART_STATEMENT_SCOPE))
 	}
 	return types.NeverType
@@ -1094,6 +1210,16 @@ func (l *Linter) lintIdent(exp *ast.Ident, ctx *context.Context) types.Type {
 			// mark table is used
 			t.IsUsed = true
 			return types.GotoType
+		} else if p, ok := ctx.Penaltyboxes[exp.Value]; ok {
+			// mark penaltybox is used
+			p.IsUsed = true
+			// Fastly treats these variables as type IDs
+			return types.IDType
+		} else if rc, ok := ctx.Ratecounters[exp.Value]; ok {
+			// mark ratecounter is used
+			rc.IsUsed = true
+			// Fastly treats these variables as type IDs
+			return types.IDType
 		} else if _, ok := ctx.Identifiers[exp.Value]; ok {
 			return types.IDType
 		}